requirements.txt
aiobotocore==2.26.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-retry==2.9.1
aioitertools==0.13.0
aiosignal==1.4.0
alabaster==1.0.0
alembic==1.17.2
altair==6.0.0
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.0
appdirs==1.4.4
asttokens==3.0.1
asyncssh==2.21.1
atpublic==7.0.0
attrs==25.4.0
babel==2.17.0
backoff==2.2.1
billiard==4.2.4
blinker==1.9.0
boto3==1.41.5
botocore==1.41.5
cachetools==6.2.2
celery==5.6.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpickle==3.1.2
colorama==0.4.6
colorlog==6.10.1
comm==0.2.3
configobj==5.0.9
contourpy==1.3.3
coverage==7.13.0
cryptography==46.0.3
cycler==0.12.1
dacite==1.6.0
dagshub==0.6.3
dagshub-annotation-converter==0.1.15
databricks-sdk==0.74.0
dataclasses-json==0.6.7
debugpy==1.8.18
decorator==5.2.1
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
docutils==0.22.3
dpath==2.2.0
dulwich==0.24.10
dvc==3.64.2
dvc-data==3.16.12
dvc-http==2.32.0
dvc-objects==5.1.2
dvc-render==1.0.2
dvc-s3==3.2.2
dvc-studio-client==0.22.0
dvc-task==0.40.2
entrypoints==0.4
exceptiongroup==1.3.1
executing==2.2.1
fastapi==0.124.2
filelock==3.20.0
flake8==7.3.0
Flask==3.1.2
flask-cors==6.0.1
flatten-dict==0.4.2
flufl.lock==8.2.0
fonttools==4.61.0
frozenlist==1.8.0
fsspec==2025.12.0
funcy==2.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.43.0
gql==4.0.0
grafana-client==5.0.1
grandalf==0.8
graphene==3.4.3
graphql-core==3.2.7
graphql-relay==3.2.0
greenlet==3.3.0
gto==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huey==2.5.5
hydra-core==1.3.2
idna==3.11
imagesize==1.4.1
importlib_metadata==8.7.0
iniconfig==2.3.0
ipykernel==7.1.0
ipython==9.8.0
ipython_pygments_lexers==1.1.1
iterative-telemetry==0.0.10
itsdangerous==2.2.0
jedi==0.19.2
jh2==5.0.10
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.7.0
jupyter_core==5.9.1
kafka-python==2.3.0
kiwisolver==1.4.9
kombu==5.6.1
lightgbm==4.6.0
llvmlite==0.46.0
lxml==6.0.2
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib==3.10.8
matplotlib-inline==0.2.1
mccabe==0.7.0
mdurl==0.1.2
mlflow==3.7.0
mlflow-skinny==3.7.0
mlflow-tracing==3.7.0
multidict==6.7.0
mypy_extensions==1.1.0
narwhals==2.13.0
nest-asyncio==1.6.0
networkx==3.6.1
niquests==3.15.2
numba==0.63.1
numpy==2.3.5
omegaconf==2.3.0
opentelemetry-api==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
optuna==4.6.0
orjson==3.11.5
packaging==25.0
pandas==2.3.3
parso==0.8.5
pathspec==0.12.1
pathvalidate==3.3.1
pillow==12.0.0
platformdirs==4.5.1
plotly==6.5.0
pluggy==1.6.0
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.2
psutil==7.1.3
pure_eval==0.2.3
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydeck==0.9.1
pydot==4.0.1
pyflakes==3.4.0
pygit2==1.19.0
Pygments==2.19.2
pygtrie==2.5.0
pyparsing==3.2.5
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
pywin32==311
PyYAML==6.0.3
pyzmq==27.1.0
qh3==1.5.6
redis==7.1.0
referencing==0.37.0
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
roman-numerals==3.1.0
rpds-py==0.30.0
rsa==4.9.1
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.15
s3fs==2025.12.0
s3transfer==0.15.0
scikit-learn==1.8.0
scipy==1.16.3
scmrepo==3.5.9
seaborn==0.13.2
semver==3.0.4
setuptools==80.9.0
shap==0.50.0
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.8.0
six==1.17.0
slicer==0.0.8
smmap==5.0.2
snowballstemmer==3.0.1
Sphinx==9.0.4
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.45
sqlparse==0.5.4
sqltrie==0.11.2
-e git+ssh://git@github.com/akashagalave/fraudguard-ai.git@442df73435a7c1a4777ec4835b848264312fe910#egg=src
stack-data==0.6.3
starlette==0.50.0
streamlit==1.52.1
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tomlkit==0.13.3
tornado==6.5.3
tqdm==4.67.1
traitlets==5.14.3
treelib==1.8.0
typer==0.20.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.6.2
urllib3-future==2.14.908
uvicorn==0.38.0
verlib2==0.3.1
vine==5.1.0
voluptuous==0.15.2
waitress==3.0.2
wassima==2.0.2
watchdog==6.0.0
wcwidth==0.2.14
Werkzeug==3.1.4
wrapt==1.17.3
xgboost==3.1.2
yarl==1.22.0
zc.lockfile==4.0
zipp==3.23.0
